# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Pay API",
        "description": "API pour l'application OM Pay - Gestion des paiements et transferts",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://om-pay-mk4h.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
        "/auth/initiate": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Initier l'inscription",
                "description": "Envoie un code OTP pour commencer le processus d'inscription",
                "operationId": "7cb5aa61d3d94bc8380474c5ab732581",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_telephone"
                                ],
                                "properties": {
                                    "numero_telephone": {
                                        "description": "Numéro de téléphone au format +2217XXXXXXXX",
                                        "type": "string",
                                        "example": "+221701234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code OTP envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code de vérification envoyé avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "abc123"
                                                },
                                                "code_otp": {
                                                    "type": "string",
                                                    "example": "123456"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Numéro de téléphone invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/verify-otp": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Vérifier le code OTP",
                "description": "Vérifie le code OTP envoyé par SMS",
                "operationId": "e27990449a29ee7e65be429d7e1b8f0f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "code"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Token reçu lors de l'initiation",
                                        "type": "string",
                                        "example": "abc123"
                                    },
                                    "code": {
                                        "description": "Code OTP à 6 chiffres",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code OTP vérifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Code OTP invalide ou expiré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code OTP invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion",
                "description": "Se connecter avec numéro de téléphone et code PIN",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_telephone",
                                    "code_pin"
                                ],
                                "properties": {
                                    "numero_telephone": {
                                        "type": "string",
                                        "example": "+221701234567"
                                    },
                                    "code_pin": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "PMOfGqub4qI1LRynqATixgXiug0PnrCjgEF7VTJqOcazKv0XrFTEnLANQJkhnBbK"
                                                },
                                                "refresh_token": {
                                                    "type": "string",
                                                    "example": "refresh_token_here"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants incorrects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Numéro ou PIN incorrect"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion",
                "description": "Se déconnecter de l'application",
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur lors de la déconnexion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/{numeroCompte}/paiement/scanner-qr": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Scanner un QR code",
                "description": "Scanne et décode un QR code de paiement marchand",
                "operationId": "4e9b0413d6ee91b924428e8d65890a86",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "donneesQR"
                                ],
                                "properties": {
                                    "donneesQR": {
                                        "description": "Données encodées dans le QR code",
                                        "type": "string",
                                        "example": "QR_CODE_DATA_HERE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "QR code scanné avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "marchand": {
                                                    "type": "object"
                                                },
                                                "montant": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "reference": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "QR code invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "QR code invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/{numeroCompte}/paiement/saisir-code": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Saisir un code de paiement",
                "description": "Saisit manuellement un code de paiement marchand",
                "operationId": "b03402b3f9fcf0bceecf7221319f98bd",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "montant"
                                ],
                                "properties": {
                                    "code": {
                                        "description": "Code de paiement fourni par le marchand",
                                        "type": "string",
                                        "example": "PAY123456"
                                    },
                                    "montant": {
                                        "description": "Montant du paiement en XOF",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code de paiement validé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "idPaiement": {
                                                    "type": "string",
                                                    "example": "OM20251111131953ABC123"
                                                },
                                                "marchand": {
                                                    "properties": {
                                                        "idMarchand": {
                                                            "type": "string"
                                                        },
                                                        "nom": {
                                                            "type": "string"
                                                        },
                                                        "logo": {
                                                            "type": "string",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "montant": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 5000
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "XOF"
                                                },
                                                "dateExpiration": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "valide": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Code de paiement invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code de paiement invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/{numeroCompte}/portefeuille/solde": {
            "get": {
                "tags": [
                    "Portefeuille"
                ],
                "summary": "Consulter le solde",
                "description": "Récupère le solde actuel du portefeuille OM Pay",
                "operationId": "49e8248ac091dc36f681590419a4ae4d",
                "parameters": [
                    {
                        "name": "numeroCompte",
                        "in": "path",
                        "description": "Numéro de compte de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7735434534"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solde récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 15000.5
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "XOF"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}